Rename `.mjs` files to `.js`
