+++ package.json
@@ -168,15 +168,15 @@
     "lint": "run-p \"lint:*\"",
     "lint:typecheck": "tsc",
     "lint:eslint": "cross-env EFF_NO_LINK_RULES=true eslint . --format friendly",
-    "lint:changelog": "node ./scripts/lint-changelog.mjs",
+    "lint:changelog": "node ./scripts/lint-changelog.js",
     "lint:prettier": "prettier . \"!test*\" --check",
     "lint:spellcheck": "cspell --no-progress --relative --dot --gitignore",
-    "lint:deps": "node ./scripts/check-deps.mjs",
+    "lint:deps": "node ./scripts/check-deps.js",
     "fix": "run-s fix:eslint fix:prettier",
     "fix:eslint": "yarn lint:eslint --fix",
     "fix:prettier": "yarn lint:prettier --write",
-    "build": "node ./scripts/build/build.mjs",
-    "build:website": "node ./scripts/build-website.mjs"
+    "build": "node ./scripts/build/build.js",
+    "build:website": "node ./scripts/build-website.js"
   },
   "c8": {
     "reporter": [
