+++ scripts/check-deps.js
