+++ scripts/generate-schema.js
