+++ scripts/lint-changelog.js
@@ -4,7 +4,7 @@ import path from "node:path";
 import fs from "node:fs";
 import { outdent } from "outdent";
 import createEsmUtils from "esm-utils";
-import { CHANGELOG_CATEGORIES } from "./utils/changelog-categories.mjs";
+import { CHANGELOG_CATEGORIES } from "./utils/changelog-categories.js";
 
 const { __dirname } = createEsmUtils(import.meta);
 const CHANGELOG_DIR = "changelog_unreleased";
