+++ scripts/utils/generate-schema.js
