+++ tsconfig.json
@@ -13,7 +13,7 @@
   "exclude": [
     "coverage/",
     "test*.*",
-    "jest.config.mjs",
+    "jest.config.js",
     // [TBD] JavaScript sources *not* affected by src/language-*:
     "src/main/ast-to-doc.js",
     "src/main/core.js",
